pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "harness-bench"
version = "0.1.0"
description = "HarnessBench: compare agentic harnesses on everyday online tasks (sister project to ClawBench)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "reacher-z and contributors" },
]
keywords = [
    "agent",
    "benchmark",
    "browser",
    "harness",
    "llm",
    "web",
    "evaluation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "clawbench-eval>=0.1.5",
    "click>=8.1",
    "platformdirs>=4.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.urls]
Homepage = "https://github.com/reacher-z/HarnessBench"
Repository = "https://github.com/reacher-z/HarnessBench"
Issues = "https://github.com/reacher-z/HarnessBench/issues"
"Sister Project" = "https://github.com/reacher-z/ClawBench"

[project.scripts]
harness-bench = "harnessbench.cli:main"
harnessbench = "harnessbench.cli:main"

# Register each built-in harness with ClawBench's plugin group so that
# `clawbench-eval` discovers them when HarnessBench is installed.
[project.entry-points."clawbench.harnesses"]
hb-openclaw    = "harnessbench.harnesses.openclaw.spec:spec"
hb-hermes      = "harnessbench.harnesses.hermes.spec:spec"
hb-claw-code   = "harnessbench.harnesses.claw_code.spec:spec"
hb-stagehand   = "harnessbench.harnesses.stagehand.spec:spec"
hb-browser-use = "harnessbench.harnesses.browser_use.spec:spec"
hb-coze-studio = "harnessbench.harnesses.coze_studio.spec:spec"

[tool.hatch.build.targets.wheel]
packages = ["src/harnessbench"]

[tool.hatch.build.targets.wheel.force-include]
"src/harnessbench/harnesses" = "harnessbench/harnesses"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
